Вынес куки в отдельный файл
