+++ src/main.js
@@ -3,3 +3,4 @@ require('./check.js');
 require('./resizer.js');
 require('./upload-stat.min.js');
 require('./upload.js');
+require('./cookie.js');
