+++ src/upload.js
@@ -7,8 +7,6 @@
  */
 'use strict';
 
-var browserCookies = require('browser-cookies');
-
 (function() {
   /** @enum {string} */
   var FileType = {
@@ -296,32 +294,6 @@ var browserCookies = require('browser-cookies');
   };
 
   fieldset.addEventListener('input', resizeFormInputHandler, true);
-
-  // Cookies
-  var formFilter = document.querySelector('#upload-filter');
-
-  var now = new Date(); // текущее время
-  var birthday = new Date();// день рождение текущего года
-  birthday.setMonth(11, 9);
-  var year = now.getFullYear(); // год текущего времени
-  var ONE_DAY = 1000 * 3600 * 24;
-
-  function cookie() {
-    if (now < birthday) {
-      birthday.setFullYear(year - 1);  // если ДР больше текущей даты, то вычитаем один из текущего года
-    }
-    return Math.round((now - birthday) / ONE_DAY); // считаем результат
-  }
-
-  formFilter.onsubmit = function() {
-
-    var cookieday = cookie();  // записываем результат функции в переменную
-    var inputFilter = document.querySelector('input[name="upload-filter"]:checked');
-
-    if (inputFilter) {
-      browserCookies.set('upload-filter', inputFilter.value, {expires: cookieday});
-    }
-  };
 
   cleanupResizer();
   updateBackground();
